+++ examples/storybook/stories/UploadMedia/index.js
@@ -1,0 +1,4 @@
+import { storiesOf } from '@storybook/react';
+import ErrorToastStory from './ErrorToasts';
+
+storiesOf('Media Upload').add('Error Toasts', ErrorToastStory);