+++ examples/storybook/stories/index.js
@@ -6,3 +6,4 @@ require('./Preview/');
 require('./Ricos/');
 require('./Toolbars/');
 require('./TestCase/');
+require('./UploadMedia/');