+++ packages/editor/web/src/RichContentEditor/Icons/CloseIcon.tsx
@@ -1,0 +1,13 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const CloseIcon = props => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" {...props}>
+    <path
+      fillRule="evenodd"
+      d="M18.77 4L20 5.23 13.23 12 20 18.77 18.77 20 12 13.23 5.23 20 4 18.77 10.769 12 4 5.23 5.23 4 12 10.769 18.77 4z"
+    />
+  </svg>
+);
+
+export default CloseIcon;