+++ packages/editor/web/src/RichContentEditor/Icons/index.ts
@@ -11,3 +11,4 @@ export { default as TitleTwoIcon } from './TitleTwoIcon';
 export { default as UnderlineIcon } from './UnderlineIcon';
 export { default as UnorderedListIcon } from './UnorderedListIcon';
 export { default as ShortcutIcon } from './ShortcutIcon';
+export { default as CloseIcon } from './CloseIcon';